backend/aiLingoNew/requirements.txt
django==4.2.11
google-generativeai
requests
python-dotenv
Pillow
django-cors-headers
gunicorn
psycopg2-binary
django-ninja